+++ src/style/components/song/_index.scss
@@ -6,8 +6,8 @@
 
 @use 'artist_widget';
 @use 'playlist_position_info';
-@use 'song';
-@use 'song_expanded';
+@use 'song_widget';
+@use 'song_expanded_widget';
 @use 'song_status';
 @use 'work_widget';
 @use 'work_link_widget';
