+++ src/style/components/song/_song_expanded_widget.scss
@@ -6,7 +6,7 @@
 
 @use '~/abstracts/sizes';
 
-.song-expanded {
+.song-expanded-widget {
   .artist-widget,
   .work-link-widget {
     align-items: center;
