+++ src/style/components/song/_song_widget.scss
@@ -12,11 +12,11 @@
 @use '~/abstracts/fonts';
 @use '~/abstracts/sizes';
 
-// `song` class:
+// `song-widget` class:
 //
 // This class stylizes a song independently of its use (as a single item or as
 // part of a listing).
-.song {
+.song-widget {
   column-gap: 0.5em;
   display: inline-flex;
   overflow: hidden;
